Taking the hsv and binary together to fin the relation
takes the binary result at 25% and HSV result at 25% and over lays them
together in an image. the image is then comapared to the trained image.
need to do the white count for the merged image. possibly do a training
that takes multiple images and merges the results to get a propper
trained image. still very slow(about 10 seconds per image).